Rename module folder ai_agents -> ampower_koda, update module references
