+++ ampower_koda/agent/api.py
@@ -7,8 +7,8 @@ import subprocess
 import frappe
 from frappe import _
 
-from ampower_ai_agents.agent.git_ops import checkout_base
-from ampower_ai_agents.agent.graph import _get_bench_env
+from ampower_koda.agent.git_ops import checkout_base
+from ampower_koda.agent.graph import _get_bench_env
 
 DOCTYPE_NAME = "AI Agent Request"
 
@@ -67,7 +67,7 @@ def start_agent(request_name: str):
     frappe.db.commit()
 
     frappe.enqueue(
-        "ampower_ai_agents.agent.executor.run_planning_phase",
+        "ampower_koda.agent.executor.run_planning_phase",
         queue="default",
         timeout=1800,
         request_name=request_name,
@@ -104,7 +104,7 @@ def execute_existing_plan(request_name: str):
     frappe.db.commit()
 
     frappe.enqueue(
-        "ampower_ai_agents.agent.executor.run_execution_phase",
+        "ampower_koda.agent.executor.run_execution_phase",
         queue="default",
         timeout=1800,
         request_name=request_name,
@@ -135,7 +135,7 @@ def approve_plan(request_name: str, edited_plan: str = None):
     frappe.db.commit()
 
     frappe.enqueue(
-        "ampower_ai_agents.agent.executor.run_execution_phase",
+        "ampower_koda.agent.executor.run_execution_phase",
         queue="default",
         timeout=1800,
         request_name=request_name,
@@ -190,7 +190,7 @@ def approve_bench(request_name: str, commands: str = None):
     frappe.db.commit()
 
     frappe.enqueue(
-        "ampower_ai_agents.agent.executor.run_bench_and_commit",
+        "ampower_koda.agent.executor.run_bench_and_commit",
         queue="default",
         timeout=1800,
         request_name=request_name,
@@ -231,7 +231,7 @@ def approve_push(request_name: str, push_branch: int = 1, create_pr: int = 1):
     frappe.db.commit()
 
     frappe.enqueue(
-        "ampower_ai_agents.agent.executor.run_deploy_phase",
+        "ampower_koda.agent.executor.run_deploy_phase",
         queue="default",
         timeout=600,
         request_name=request_name,
