+++ ampower_koda/ampower_koda/doctype/ai_agent_prompt_configuration/ai_agent_prompt_configuration.json
@@ -26,7 +26,7 @@
         }
     ],
     "istable": 1,
-    "module": "AI Agents",
+    "module": "Ampower Koda",
     "name": "AI Agent Prompt Configuration",
     "owner": "Administrator"
 }
