+++ ampower_koda/ampower_koda/doctype/ai_agent_request/ai_agent_request.json
@@ -334,7 +334,7 @@
     "links": [],
     "modified": "2026-03-12 10:00:00.000000",
     "modified_by": "Administrator",
-    "module": "AI Agents",
+    "module": "Ampower Koda",
     "name": "AI Agent Request",
     "naming_rule": "Expression",
     "owner": "Administrator",
