+++ ampower_koda/ampower_koda/doctype/ai_agents_settings/ai_agents_settings.json
@@ -109,7 +109,7 @@
     "links": [],
     "modified": "2026-03-27 12:00:00.000000",
     "modified_by": "Administrator",
-    "module": "AI Agents",
+    "module": "Ampower Koda",
     "name": "AI Agents Settings",
     "owner": "Administrator",
     "permissions": [
