+++ ampower_koda/ampower_koda/page/ai_agents_help/ai_agents_help.json
@@ -6,7 +6,7 @@
     "idx": 0,
     "modified": "2026-03-12 10:00:00.000000",
     "modified_by": "Administrator",
-    "module": "AI Agents",
+    "module": "Ampower Koda",
     "name": "ai-agents-help",
     "owner": "Administrator",
     "page_name": "ai-agents-help",
