+++ ampower_koda/ampower_koda/workspace/ai_agents/ai_agents.json
@@ -45,7 +45,7 @@
     ],
     "modified": "2026-03-05 12:00:00.000000",
     "modified_by": "Administrator",
-    "module": "AI Agents",
+    "module": "Ampower Koda",
     "name": "AI Agents",
     "number_cards": [],
     "owner": "Administrator",
