+++ ampower_koda/hooks.py
@@ -1,5 +1,5 @@
-app_name = "ampower_ai_agents"
-app_title = "AmPower AI Agents"
+app_name = "ampower_koda"
+app_title = "AmPower Koda"
 app_publisher = "Ambibuzz Technologies LLP"
 app_description = "AI Coding Agent for Frappe apps"
 app_email = "buzz@ambibuzz.com"
@@ -7,5 +7,5 @@ app_license = "MIT"
 required_apps = ["frappe"]
 
 doctype_js = {
-    "AI Agent Request": "ai_agents/doctype/ai_agent_request/ai_agent_request.js",
+    "AI Agent Request": "ampower_koda/doctype/ai_agent_request/ai_agent_request.js",
 }
